+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Use CI friendly Maven version.
 * Release build on CI tag using Github action
 * Updated libs.
+* Login Dialog for multiuser mode.
 
 ## Fixes
 * \#67 Aggregated reports work also with sorting.